Condense helper classes and functions into single utils module
